Update patch set 3
Patch Set 3:

> > > This sets the variables without the vendor prefix again, thanks.
> > > Unfortunately BT still crashes for some reason.  I need time to
> > > debug this.
> > 
> > That's probably related to mac adress. You can try with something like this: https://github.com/LineageOS/android_device_cyanogen_msm8916-common/commit/918dff7e77b93184e3b19941ace4cc81c878639f
> 
> Thanks, but that doesn't help either, unfortunately.

Actually, after applying this patch and restarting the device it turns out that the *only* wc_transport property we have is

  [wc_transport.stack_bdaddr]: [38:76:a0:01:7c:a8]

All the other properties are missing.  When inspecting the dmesg output
it turns out that "open(/dev/smd3 fails with error -517".

I assume there's some change missing but I don't see what that may be.

Corinna

Patch-set: 3